change to 16bit colors

+++ Core/Inc/hub75_ospi.h
@@ -84,7 +84,7 @@ void HUB75_StopDrawing(void);
 
 void HUB75_SwapDisplayFrame(void);
 
-void HUB75_CopyFrame(uint8_t *frame, uint16_t size);
+void HUB75_CopyFrame(uint16_t *frame, uint16_t size);
 
 /**
  * @brief  Write a 1-bit-per-channel pixel to the software framebuffer.
